Refactor to use DashboardRouter
